==> notebooks/03_Database_Population.ipynb <==
from sqlalchemy import create_engine
import pandas as pd

username = 'postgres'  
password = 'XXXXXXXXX'  
database = 'airlines' 
host = 'localhost' 
port = 5432

url = f"postgresql://{username}:{password}@{host}:{port}/{database}"

engine = create_engine(url)

print("Zmienna `url` i `engine` zostały zdefiniowane.")

paths = {
    "aircraft": r"C:\Users\aga_c\Downloads\Pliki_do_Portfolio_Lab (1)\notebooks\aircraft.csv",
    "airport_list": r"C:\Users\aga_c\Downloads\Pliki_do_Portfolio_Lab (1)\notebooks\airport_data.csv",
    "airport_weather": r"C:\Users\aga_c\Downloads\Pliki_do_Portfolio_Lab (1)\notebooks\airport_weather.csv",
    "flight": r"C:\Users\aga_c\Downloads\Pliki_do_Portfolio_Lab (1)\notebooks\flight.csv",
}

dataframes = {name: pd.read_csv(path) for name, path in paths.items()}

def export_table_to_db(df, table_name, engine):
 

    if 'id' not in df.columns:
        df.insert(0, 'id', range(1, len(df) + 1))
    

    df.to_sql(table_name, con=engine, if_exists='replace', index=False)
    print(f"Tabela '{table_name}' została wyeksportowana do bazy danych.")

connection_string = f"postgresql://{username}:{password}@{host}:{port}/{database}"
engine = create_engine(connection_string)

table_names = {
    "aircraft": "aircraft",
    "airport_list": "airport_list",
    "airport_weather": "airport_weather",
    "flight": "flight",
}

for df_name, table_name in table_names.items():
    export_table_to_db(dataframes[df_name], table_name, engine)
