==> notebooks/05_Report Schema_Design.ipynb <==
import psycopg2

username = 'postgres'  
password = 'XXXXXXXXXXXXXX'  
database = 'airlines'
host = 'localhost'  
port = 5432

try:
    connection = psycopg2.connect(
        user=username,
        password=password,
        database=database,
        host=host,
        port=port
    )
    cursor = connection.cursor()
    print("Połączenie z bazą danych zostało nawiązane.")
except Exception as error:
    print(f"Błąd podczas łączenia z bazą danych: {error}")
    connection = None  


file_path = r"C:\Users\aga_c\Downloads\Pliki_do_Portfolio_Lab (1)\sql\reporting.sql"

try:
    with open(file_path, 'r', encoding='utf-8') as file:
        sql_script = file.read()
    
    cursor.execute(sql_script)
    con.commit() 
    print("Skrypt SQL został pomyślnie wykonany.")
except Exception as e:
    print(f"Wystąpił błąd: {e}")


if connection is not None:
    cursor.close()
    connection.close()
    print("Połączenie z bazą danych zostało zamknięte.")
